orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
httpx==0.27.0
python-multipart==0.0.9
